Revert rmsd.py and test_rmsd.py Add parallel.py (with prmsdwrapper) and test_parallel.py
